+++ packages/Webkul/UI/src/DataGrid/Traits/ProvideTabFilters.php
@@ -3,6 +3,7 @@
 namespace Webkul\UI\DataGrid\Traits;
 
 use Carbon\Carbon;
+use Illuminate\Support\Facades\DB;
 
 trait ProvideTabFilters
 {
@@ -27,6 +28,34 @@ trait ProvideTabFilters
      */
     public function prepareTabFilters()
     {
+    }
+
+    /**
+     * Add tab filter.
+     *
+     * @param  array  $filterConfig
+     * @return void
+     */
+    public function addTabFilter($filterConfig)
+    {
+        if (($filterConfig['value_type'] ?? false) == "lookup") {
+            $values = app($filterConfig['repositoryClass'])
+                ->get(['name', 'code as key', DB::raw("false as isActive")])
+                ->prepend([
+                    'isActive'  => true,
+                    'key'       => 'all',
+                    'name'      => trans('admin::app.datagrid.all'),
+                ])
+                ->toArray();
+
+            $filterConfig['values'] = $values;
+        } else {
+            foreach ($filterConfig['values'] as $valueIndex => $value) {
+                $filterConfig['values'][$valueIndex]['name'] = trans($filterConfig['values'][$valueIndex]['name']);
+            }
+        }
+
+        $this->tabFilters[] = $filterConfig;
     }
 
     /**
